% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjdepot-package.R
\docType{package}
\name{tjdepot-package}
\alias{tjdepot}
\alias{tjdepot-package}
\title{tjdepot: Analysis Depot for TJ Palanca}
\description{
Analysis depot for TJ Palanca
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjpalanca/tjdepot}
  \item Report bugs at \url{https://github.com/tjpalanca/tjdepot/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
